lucetc: move pointer consts to their own place
