+++ lucetc/src/new/function.rs
@@ -1,5 +1,5 @@
 use super::runtime::RuntimeFunc;
-use crate::compiler::entity::{NATIVE_POINTER, POINTER_SIZE};
+use crate::new::pointer::{NATIVE_POINTER, POINTER_SIZE};
 use crate::new::decls::ModuleDecls;
 use cranelift_codegen::cursor::FuncCursor;
 use cranelift_codegen::ir::{self, InstBuilder};
