+++ lucetc/src/new/pointer.rs
@@ -1,0 +1,5 @@
+
+use cranelift_codegen::ir;
+
+pub const POINTER_SIZE: usize = 8;
+pub const NATIVE_POINTER: ir::Type = ir::types::I64;
